Adds redis setup to bin/setup (sorry for the macOS assumption)
In getting everything working on a separate system, I figured having
redis be a part of the installation was probably worthwhile.
Making the assumption that you're on macOS and have homebrew already
installed (hopefully if that's not the case, you can update the setup
script to make it work for your system in addition to my default
assumption).